Added thread speed testing

+++ modules/fps.py
@@ -9,12 +9,12 @@ class FPS:
 
     def __init__(self):
 
-        self._fps = 0
-        self._processing_time = 0
+        self._fps = 0.0
+        self._processing_time = 0.0
         self._processing_times = []
 
-        self._last_call_time = 0
-        self._update_time = 0
+        self._last_call_time = 0.0
+        self._update_time = 0.0
 
     def tick(self):
         self._processing_time = time.time() - self._last_call_time
@@ -23,7 +23,7 @@ class FPS:
         self._processing_times.append(self._processing_time)
 
         if (time.time() - self._update_time) > UPDATE_DELAY:
-            self._fps = int(1 / numpy.mean(self._processing_times))
+            self._fps = 1 / numpy.mean(self._processing_times)
             self._processing_times = []
             self._update_time = time.time()
